requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
pydantic
openai
python-dotenv
